12-01-25 Monday 5pm - Updates to MyRunsClient.tsx and DashboardClient.tsx
For my-runs page - Under each run, it's displaying the TITLE of the highlight video but instead if the HighlightID/jobId

For the dashboard page - the Open button is renamed as the Play button.
Left-clicking the Play button to watch the highlight video in the Highlight Videos gallery.
Right-clicking the button & open a new tab to watch the video in the HoopTuber Video Player.